# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gym CRM API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/trainers/{username}": {
      "get": {
        "tags": [
          "Trainer Management"
        ],
        "summary": "Get trainer profile by username",
        "description": "Returns trainer profile details including specialization, assigned trainees (if applicable), and active status.",
        "operationId": "getProfile",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainer whose profile will be retrieved",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "trainer_jane"
          }
        ],
        "responses": {
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainer not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      },
      "put": {
        "tags": [
          "Trainer Management"
        ],
        "summary": "Update trainer profile",
        "description": "Updates trainer profile details such as personal information and specialization (depending on request model).",
        "operationId": "updateProfile",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainer whose profile will be updated",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "trainer_jane"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainer not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
    "/api/v1/trainees/{username}": {
      "get": {
        "tags": [
          "Trainee Management"
        ],
        "summary": "Get trainee profile by username",
        "description": "Returns trainee profile details including personal info, assigned trainers and active status.",
        "operationId": "getProfile_1",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainee whose profile will be retrieved",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "john_doe"
          }
        ],
        "responses": {
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainee not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      },
      "put": {
        "tags": [
          "Trainee Management"
        ],
        "summary": "Update trainee profile",
        "description": "Updates trainee profile details. The request may include personal details and address fields.",
        "operationId": "updateProfile_1",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainee whose profile will be updated",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "john_doe"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraineeProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainee not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Trainee Management"
        ],
        "summary": "Delete trainee profile",
        "description": "Deletes the trainee profile and related associations.",
        "operationId": "deleteProfile",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainee whose profile will be deleted",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "john_doe"
          }
        ],
        "responses": {
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainee not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "200": {
            "description": "Profile deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
    "/api/v1/trainers": {
      "post": {
        "tags": [
          "Trainer Management"
        ],
        "summary": "Register a new trainer",
        "description": "Creates a new trainer profile and returns registration details.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "409": {
            "description": "Conflict (e.g., username already exists)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "200": {
            "description": "Trainer registered successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          }
        ]
      }
    },
...
    "/api/v1/trainers/{username}/active-status": {
      "patch": {
        "tags": [
          "Trainer Management"
        ],
        "summary": "Set trainer active status",
        "description": "Activates or deactivates the trainer profile.",
        "operationId": "setActiveStatus",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "trainer_jane"
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "New active status",
            "required": true,
            "schema": {
              "type": "boolean"
            },
            "example": true
          }
        ],
        "responses": {
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "200": {
            "description": "Active status updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainer not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
    "/api/v1/trainees/{username}/active-status": {
      "patch": {
        "tags": [
          "Trainee Management"
        ],
        "summary": "Set trainee active status",
        "description": "Activates or deactivates the trainee profile.",
        "operationId": "setActiveStatus_1",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainee",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "john_doe"
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "New active status",
            "required": true,
            "schema": {
              "type": "boolean"
            },
            "example": true
          }
        ],
        "responses": {
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainee not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "200": {
            "description": "Active status updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
    "/api/v1/trainers/{username}/trainings": {
      "get": {
        "tags": [
          "Trainer Management"
        ],
        "summary": "Get trainer trainings with filters",
        "description": "Returns training sessions for the specified trainer. Optional filters may include date range and trainee-related fields (depending on filter DTO).",
        "operationId": "getTrainings",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "trainer_jane"
          },
          {
            "name": "filters",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TrainerTrainingFilterRequest"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "200": {
            "description": "Trainings retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainer not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
    "/api/v1/trainees/{username}/trainings": {
      "get": {
        "tags": [
          "Trainee Management"
        ],
        "summary": "Get trainee trainings with filters",
        "description": "Returns trainee training sessions. Optional filters may include date range and trainer-related fields (depending on filter DTO).",
        "operationId": "getTrainings_1",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the trainee",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "john_doe"
          },
          {
            "name": "filters",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TraineeTrainingFilterRequest"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Validation error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "200": {
            "description": "Trainings retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trainee not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStandardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "customAuth": []
          },
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "TrainerProfileUpdateRequest": {
        "required": [
          "firstName",
          "isActive",
          "lastName",
          "specializationId"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 64,
            "minLength": 2,
            "type": "string",
            "description": "Trainer's first name.",
            "example": "Jane"
          },
          "lastName": {
            "maxLength": 64,
            "minLength": 2,
            "type": "string",
            "description": "Trainer's last name.",
            "example": "Smith"
          },
          "specializationId": {
            "type": "integer",
            "description": "Identifier of the trainer's specialization (training type).",
            "format": "int64",
            "example": 2
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates whether the trainer profile is active.",
            "example": true
          }
        },
        "description": "Request body for updating a trainer's profile information."
      },
...
      "ApiStandardResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TrainerProfileUpdateResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "description": "Standard API response wrapper. Contains either data on success or error details on failure."
      },
...
      "TraineeProfileUpdateRequest": {
        "required": [
          "firstName",
          "isActive",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 64,
            "minLength": 2,
            "type": "string",
            "description": "Trainee's first name.",
            "example": "John"
          },
          "lastName": {
            "maxLength": 64,
            "minLength": 2,
            "type": "string",
            "description": "Trainee's last name.",
            "example": "Doe"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Trainee's date of birth (ISO-8601, yyyy-MM-dd).",
            "format": "date",
            "example": "1990-01-01"
          },
          "address": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "description": "Trainee's address.",
            "example": "123 Main St, Anytown, USA"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the trainee profile is active.",
            "example": true
          }
        },
        "description": "Request body for updating a trainee profile, including personal details and active status."
      },
...
      "TrainerRegisterRequest": {
        "required": [
          "firstName",
          "lastName",
          "specializationId"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 64,
            "minLength": 2,
            "type": "string",
            "description": "Trainer's first name.",
            "example": "Jane"
          },
          "lastName": {
            "maxLength": 64,
            "minLength": 2,
            "type": "string",
            "description": "Trainer's last name.",
            "example": "Smith"
          },
          "specializationId": {
            "type": "integer",
            "description": "Identifier of the trainer's specialization (training type).",
            "format": "int64",
            "example": 1
          }
        },
        "description": "Request body for registering a new trainer, including personal details and specialization."
      },
...
      "TrainerTrainingFilterRequest": {
        "type": "object",
        "properties": {
          "periodFrom": {
            "type": "string",
            "description": "Start date of the training period (inclusive). ISO-8601 format (yyyy-MM-dd).",
            "format": "date",
            "example": "2024-01-01"
          },
          "periodTo": {
            "type": "string",
            "description": "End date of the training period (inclusive). ISO-8601 format (yyyy-MM-dd).",
            "format": "date",
            "example": "2024-12-31"
          },
          "traineeName": {
            "type": "string",
            "description": "Full name of the trainee used to filter training sessions.",
            "example": "John Doe"
          }
        },
        "description": "Query parameters used to filter a trainer's training sessions by date range and trainee name."
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}